Making a change on a merged PR
Making a change on a merged PR as I haven't deleted the branch yet. Want to see how it works (and also to see if the comment will show up if I merge normally or if it will only show up if I merge squash. I think it will show up either way but I guess we'll find out.